requirements.txt
torch
torchvision
numpy
pandas
pillow
matplotlib
tqdm
scikit-learn
opencv-python
seaborn
